app/requirements.txt
chardet
fastapi
numpy
pandas
pydantic
starlette
uvicorn
python-multipart
pipreqs
pytest
pyyaml
